Moving copy indexes/constraints functions to command
