+++ usaspending_api/etl/management/commands/copy_table_metadata.py
@@ -1,14 +1,63 @@
 import logging
+import re
 from django import db
 from django.core.management.base import BaseCommand
 
+from usaspending_api.etl.broker_etl_helpers import dictfetchall
 from usaspending_api.etl.management.commands.combine_transaction_search_chunks import Command as CTSC_Command
 from usaspending_api.database_scripts.matview_generator.chunked_matview_sql_generator import (
-    make_copy_indexes,
-    make_copy_constraints,
+    make_read_constraints,
+    make_read_indexes,
 )
 
 logger = logging.getLogger("script")
+
+
+def make_copy_constraints(cursor, source_table, dest_table, drop_foreign_keys=False):
+    # read the existing indexes
+    cursor.execute(make_read_constraints(source_table)[0])
+    src_constrs = dictfetchall(cursor)
+
+    # build the destination index sql
+    dest_constr_sql = []
+    for constr_dict in src_constrs:
+        create_constr_name = constr_dict["conname"]
+        create_constr_content = constr_dict["pg_get_constraintdef"]
+        if "FOREIGN KEY" in create_constr_content and drop_foreign_keys:
+            continue
+        dest_constr_sql.append(
+            f"ALTER TABLE {dest_table} ADD CONSTRAINT {create_constr_name}_temp" f" {create_constr_content}"
+        )
+    return dest_constr_sql
+
+
+def make_copy_indexes(cursor, source_table, dest_table):
+    # read the existing indexes of source table
+    cursor.execute(make_read_indexes(source_table)[0])
+    src_indexes = dictfetchall(cursor)
+
+    # reading the existing indexes of destination table (to not duplicate anything)
+    cursor.execute(make_read_indexes(dest_table)[0])
+    dest_indexes = [ix_dict["indexname"] for ix_dict in dictfetchall(cursor)]
+
+    # build the destination index sql
+    dest_ix_sql = []
+    for ix_dict in src_indexes:
+        ix_name = ix_dict["indexname"]
+        dest_ix_name = f"{ix_name}_temp"
+        if dest_ix_name in dest_indexes:
+            logger.info(f"Index {dest_ix_name} already in {dest_table}. Skipping.")
+            continue
+
+        create_ix_sql = ix_dict["indexdef"]
+        ix_regex = r"CREATE\s.*INDEX\s\S+\sON\s(\S+)\s.*"
+        # this *should* match source_table, but can get funky with/without the schema included and regex
+        # for example, a table 'x' in the public schema could be provided and the string will include `public.x'
+        src_table = re.findall(ix_regex, create_ix_sql)[0]
+        create_ix_sql = create_ix_sql.replace(ix_name, dest_ix_name)
+        create_ix_sql = create_ix_sql.replace(src_table, dest_table)
+        dest_ix_sql.append(create_ix_sql)
+    return dest_ix_sql
 
 
 class Command(BaseCommand):
